fix tubro back button handling when window.history.length is > 0 on page load
If you navigate to a turbo site from an existing browser session, the window. history.length is > 0. (This is what happens when one reuses a browser window.)
If window.history.length > 0 when turbo initially loads then turbo’s support for the back button is broken.

Changing the pageCache from an array to a hash fixes this problem. 