Fix disappearing progress bar when using transition cache
The progress bar will disappear when you visit 2 pages which load between half speed and speed (150 ms and 300 ms).

When using transition cache ProgressBar#done and consecutively ProgressBar#_reset are called twice, once when the page is loaded from the cache, once after the replacement page is fetched. ProgressBar#_reset stores the current opacity and starts 2 timers, one with half speed which sets the opacity to 0 and one with the normal speed which restores the opacity. If you manage to load a page between these timeouts (150 ms and 300 ms) the opacity will be 0 when #_reset is called for the second time, when the second timer expires it will set the opacity to 0 forever.

This will be fixed by making sure ProgressBar#done is called only once.