Improve progress bar animation when interrupted
If you click a link while the progress bar animation is still running the animation doesn't look so smooth, it doesn't fade out nicely either.
Now when the progress bar is started while it's already running it will reset the animation and any running timers, force a reflow and then start over.