Fix page caching and lifecycle events
Fix #551.

Partial replacement broke the page cache by caching body.outerHTML (losing all DOM
state in the process) instead of the body element itself.

This commit brings back the old behavior, with the following gotcha: when a partial
replacement is performed, we remove the current page from the cache, since the body
element will not change and there is no simple way for us to bring back the changed
nodes.

Additionally:

- page:after-remove on body elements now triggers on cache eviction (since we don't
  want to lose DOM state)
- to avoid triggering page:load on the same body elements more than once, a partial
  replacement now triggers the page:partial-load event